fix: use np.int32 for eval_hermite Windows

+++ gbasis/evals/_deriv.py
@@ -132,7 +132,7 @@ def _eval_deriv_contractions(coords, orders, center, angmom_comps, alphas, prim_
         coeffs = np.array(coeffs, dtype=np.float64)
         _herm_arg = np.array(alphas**0.5 * nonzero_coords, dtype=np.float64)
         hermite = np.sum(
-            coeffs * eval_hermite(indices_herm.astype(int), _herm_arg), axis=0
+            coeffs * eval_hermite(indices_herm.astype(np.int32), _herm_arg), axis=0
         )
         hermite = np.prod(hermite, axis=1)
